+++ breeze-dream/src/main/kotlin/com/windea/breezeframework/dream/enums/property/ImpactAttribute.kt
@@ -1,25 +1,25 @@
 package com.windea.breezeframework.dream.enums.property
 
 import com.windea.breezeframework.core.annotations.messages.*
-import com.windea.breezeframework.core.consts.Locales.SimpleChinese
+import com.windea.breezeframework.core.consts.Locales.simpleChinese
 
 /**作用属性。*/
-@Name("作用属性", SimpleChinese)
+@Name("作用属性", simpleChinese)
 enum class ImpactAttribute {
-	@Name("物理", SimpleChinese)
+	@Name("物理", simpleChinese)
 	Physics,
-	@Name("化学", SimpleChinese)
+	@Name("化学", simpleChinese)
 	Chemical,
-	@Name("灵魂", SimpleChinese)
+	@Name("灵魂", simpleChinese)
 	Soul,
-	@Name("灵力", SimpleChinese)
+	@Name("灵力", simpleChinese)
 	Spirit,
-	@Name("风息", SimpleChinese)
+	@Name("风息", simpleChinese)
 	Wind,
-	@Name("冷气", SimpleChinese)
+	@Name("冷气", simpleChinese)
 	Cold,
-	@Name("火炎", SimpleChinese)
+	@Name("火炎", simpleChinese)
 	Fire,
-	@Name("电磁", SimpleChinese)
+	@Name("电磁", simpleChinese)
 	Electric
 }