+++ breeze-dsl/src/test/kotlin/com/windea/breezeframework/dsl/text/CriticMarkupDslTest.kt
@@ -1,0 +1,35 @@
+package com.windea.breezeframework.dsl.text
+
+import com.windea.breezeframework.dsl.text.CriticMarkupInlineBuilder.cmAppend
+import com.windea.breezeframework.dsl.text.CriticMarkupInlineBuilder.cmHighlight
+import org.junit.*
+
+class CriticMarkupDslTest {
+	@Test
+	fun test1() {
+		criticMarkup {
+			textBlock("""
+				Hello, ${cmAppend("world")}!
+				Nice to ${cmHighlight("meet")} you!
+			""".trimIndent())
+		}.also { println(it) }
+	}
+	
+	@Test
+	fun test2() {
+		criticMarkup {
+			+"""
+				Hello, ${cmAppend("world")}!
+				Nice to ${cmHighlight("meet")} you!
+			""".trimIndent()
+		}.also { println(it) }
+	}
+	
+	@Test
+	fun test3() {
+		"""
+			Hello, ${cmAppend("world")}!
+			Nice to ${cmHighlight("meet")} you!
+		""".trimIndent().also { println(it) }
+	}
+}